+++ lib/css.mli
@@ -6968,6 +6968,7 @@ type list_style_type = Properties.list_style_type =
   | Trad_chinese_informal
   | Trad_chinese_formal
   | Ethiopic_numeric
+  | Name of string  (** A case-sensitive custom counter-style name. *)
   | String of string
   | Symbols of symbols_type option * list_style_symbol list
   | Inherit